requirements.txt
google-cloud-bigquery
Faker==25.4.0